chore(pkg,agent): separate interfaces and structure from client's implementation
