+++ pkg/api/client/client.go
@@ -1,0 +1,36 @@
+package client
+
+import (
+	"context"
+
+	resty "github.com/go-resty/resty/v2"
+	"github.com/shellhub-io/shellhub/pkg/models"
+	"github.com/shellhub-io/shellhub/pkg/revdial"
+	"github.com/sirupsen/logrus"
+)
+
+type commonAPI interface {
+	ListDevices() ([]models.Device, error)
+	GetDevice(uid string) (*models.Device, error)
+}
+
+type publicAPI interface {
+	GetInfo(agentVersion string) (*models.Info, error)
+	Endpoints() (*models.Endpoints, error)
+	AuthDevice(req *models.DeviceAuthRequest) (*models.DeviceAuthResponse, error)
+	NewReverseListener(ctx context.Context, token string) (*revdial.Listener, error)
+	AuthPublicKey(req *models.PublicKeyAuthRequest, token string) (*models.PublicKeyAuthResponse, error)
+}
+
+type Client interface {
+	commonAPI
+	publicAPI
+}
+
+type client struct {
+	scheme string
+	host   string
+	port   int
+	http   *resty.Client
+	logger *logrus.Logger
+}
